fix: 修复🔵提示级问题 T5/T6/T11/T13/AT2/AT3/AT4
- T5: publishStreamCancelledFinal 格式对齐 Python（type=event + is_complete + event_type=chat.final）
- T6: hasActiveTask 从 *bool 改为 bool 值类型，默认 false，不再序列化 null
- T11: SandboxGatewayConfig.TimeoutSeconds 从 float64 改为 int，对齐 Python
- T13: HandleHeartbeat 空 params 时用 json.Marshal 替代字符串拼接，避免注入风险
- AT2: TeamAgentSpec ModelPool 改为 []ModelPoolEntry, ModelRouter 改为 *ModelRouterConfig
- AT3: TeamSpec ModelPool 改为 []ModelPoolEntry
- AT4: TeamRuntimeContext MessagerConfig 改为 *MessagerTransportConfig, DBConfig 改为 DatabaseConfig
- AT1: 暂不改（影响范围大）

+++ internal/agentcore/sys_operation/config.go
@@ -77,7 +77,7 @@ type SandboxGatewayConfig struct {
 	// 生产中 100% 使用 PreDeployLauncherConfig，Go 用具体类型代替 Python Union。
 	LauncherConfig *PreDeployLauncherConfig `yaml:"launcher_config" json:"launcher_config"`
 	// TimeoutSeconds 统一超时时间（秒），包含请求+就绪检测
-	TimeoutSeconds float64 `yaml:"timeout_seconds" json:"timeout_seconds"`
+	TimeoutSeconds int `yaml:"timeout_seconds" json:"timeout_seconds"`
 	// AuthHeaders 认证 HTTP 头
 	AuthHeaders map[string]string `yaml:"auth_headers" json:"auth_headers,omitempty"`
 	// AuthQueryParams 认证查询参数
@@ -193,7 +193,7 @@ func NewSandboxGatewayConfig() *SandboxGatewayConfig {
 	return &SandboxGatewayConfig{
 		Isolation:      NewSandboxIsolationConfig(),
 		LauncherConfig: NewPreDeployLauncherConfig(""),
-		TimeoutSeconds: 30.0,
+		TimeoutSeconds: 30,
 	}
 }
